+++ src/Domain/Serializer/Serializable.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Ob\Hex\Domain\Serializer;
+
+interface Serializable
+{
+    /**
+     * @return array
+     */
+    public function serialize();
+
+    /**
+     * @return mixed
+     */
+    public static function unserialize(array $data);
+}