+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All Notable changes to `Hex` will be documented in this file.
 
 ### Added
 
+- `JsonSerializer` to encode entities as a json string, or decode a string back to an object
+- `Serializable` interface that entities can implement to be, guess what, serializable
 - `MeetingRoom` entity
 - `Reservation` entity
 - `Email` value object